fix: Seat 의 SeatNumber Long타입으로 변경

+++ src/main/kotlin/io/hhplus/concert/domain/seat/Seat.kt
@@ -13,7 +13,7 @@ class Seat(
     val id: Long = 0L,
 
     @Column(name = "number")
-    val seatNumber: Int = 0,
+    val seatNumber: Long = 0,
 
     @ManyToOne
     @JoinColumn(name = "schedule_id")
